Adding group control per logged in

+++ resources/group.py
@@ -5,7 +5,7 @@ from flask_smorest import Blueprint, abort
 
 from sqlalchemy.exc import SQLAlchemyError, IntegrityError
 from models import GroupModel
-from flask_jwt_extended import get_jwt, jwt_required
+from flask_jwt_extended import get_jwt, jwt_required, get_jwt_identity
 from db import db
 from schemas import GroupSchema
 
@@ -19,11 +19,19 @@ class Group(MethodView):
     @blp.response(200, GroupSchema)
     def get(self, group_id):
         group = GroupModel.query.get_or_404(group_id)
+        current_user_id = get_jwt_identity()
+        if group.user_id != current_user_id:
+            abort(403, message="You are not authorized to view this group.")
+
         return group
 
     @jwt_required()
     def delete(self, group_id):
         group = GroupModel.query.get_or_404(group_id)
+
+        current_user_id = get_jwt_identity()
+        if group.user_id != current_user_id:
+            abort(403, message="You are not authorized to delete this group.")
         db.session.delete(group)
         db.session.commit()
         return {"message": "group deleted"}
@@ -34,13 +42,21 @@ class GroupList(MethodView):
     @jwt_required()
     @blp.response(200, GroupSchema(many=True))
     def get(self):
-        return GroupModel.query.all()
+        user_id = get_jwt_identity()
+
+        # Filter groups by the user_id to get only the groups belonging to the logged-in user
+        groups = GroupModel.query.filter_by(user_id=user_id).all()
+        return groups
+
 
     @jwt_required()
     @blp.arguments(GroupSchema)
     @blp.response(201, GroupSchema)
     def post(self,
              group_data):  # the method only has this other param because marshmallow is feeding it with the blp.schema
+
+        group_data['user_id'] = get_jwt_identity()  # Automatically link to the logged-in user
+
         item = GroupModel(**group_data)  # just passing the arguments to create the item
 
         try:
